Worked on some open issues from last commit:
Issue 1: Tried implementing conditional expressions as xpath expressions, but ran into the problem that the .net compact framework does not have the required support for xpath, will be looking at other options next.

Issue 2: elseif element has been added to the xsd and implemented in the OrchestrationManager (It would be good to do some further testing, but it would be nice to have some more useful conditional expressions for that first).

Issue 3: solved, movenext is now called at the end of the TaskletClosing event. (After the output of the tasklet has been read.)


git-svn-id: svn+ssh://ssh.itu.dk/import/stud/share/projects/bpl/svn/trunk/impl@2144 d463407e-c81c-0410-9dde-f82ba9d86a8d